backend/requirements.txt
flask
psycopg2-binary
python-dotenv
requests
flask-cors
flask-jwt-extended